first connection to the server from site

+++ backend/ez-tasker/src/main/java/com/saar/eztasker/srv/UserService.java
@@ -25,7 +25,7 @@ public class UserService {
         return userRepository.save(user);
     }
 
-    public User getUserById(Integer id) {
+    public User getUserById(long id) {
         return userRepository.findById(id);
     }
